[WPE][GTK] Fix several warnings when building for ARMv7 (32-bits) https://bugs.webkit.org/show_bug.cgi?id=247873
Reviewed by Michael Catanzaro.

This fixes several warnings that appear when building for ARMv7 (32-bits):
 - deprecated-enum-enum-conversion
 - unused-variable
 - unused-result
 - overflow
 - sign-compare

It also disables two warnings for old versions of GCC
(old versions give false positives).

* PerformanceTests/MallocBench/MallocBench/stress_aligned.cpp:
(benchmark_stress_aligned):
* Source/JavaScriptCore/assembler/ARMv7Assembler.h:
(JSC::ARMv7Assembler::revertJumpTo_movT3movtcmpT2):
(JSC::ARMv7Assembler::ARMInstructionFormatter::twoWordOp12Reg4FourFours):
(JSC::ARMv7Assembler::ARMInstructionFormatter::twoWordOp12Reg4Reg4Imm12):
(JSC::ARMv7Assembler::ARMInstructionFormatter::twoWordOp12Reg4Reg4Reg4Imm8):
(JSC::ARMv7Assembler::ARMInstructionFormatter::twoWordOp12Reg40Imm3Reg4Imm20Imm5):
* Source/WebCore/accessibility/atspi/AccessibilityObjectTextAtspi.cpp:
(WebCore::AccessibilityObjectAtspi::characterIndex const):
(WebCore::AccessibilityObjectAtspi::selectionChanged):
* Source/WebCore/platform/graphics/filters/software/FECompositeSoftwareApplier.cpp:
(WebCore::FECompositeSoftwareApplier::applyArithmetic const):
* Source/WebCore/platform/graphics/gbm/DMABufReleaseFlag.h:
(WebCore::DMABufReleaseFlag::release):
* Source/cmake/WebKitCompilerFlags.cmake:

Canonical link: https://commits.webkit.org/256792@main

+++ PerformanceTests/MallocBench/MallocBench/stress_aligned.cpp
@@ -148,7 +148,7 @@ void benchmark_stress_aligned(CommandLine&)
     
     srandom(1); // For consistency between runs.
 
-    size_t limit = 0x00001ffffffffffful;
+    uint64_t limit = 0x00001ffffffffffful;
     
     for (size_t size = 0; size < limit; size = std::max(size, sizeof(void*)) * 2) {
         for (size_t alignment = sizeof(void*); alignment < limit; alignment *= 2) {
